template/requirements.txt
torch==2.7.0
torchvision==0.22.0
open_clip_torch==2.32.0
Pillow==11.2.1
numpy==2.2.6
